Add files via upload
Swiggy's objective is to integrate with restaurant partners and provide an optimal experience to our customers. Therefore, Swiggy has a no minimum order policy and is currently the only service in the market to offer live tracking of orders. Here we used programming languages ​​like PHP, MySql, HTML, CSS and JavaScript. This is not a real Swiggy website. This is just a copy of the Swiggy website.